refactor of screen dismiss
